Get internationalizated string using the preference Language Code
